[axl3] change of ANYTOKEN logic in TILL parser generator
before ANYTOKEN productions were added to empty slots only, therefore they ALWAYS lost to regular productions. now, they are added even if parse table slot is not empty as a conflict (and resolved later via lookahead DFA process as other conflicts). this is done so it could be possible to choose ANYTOKEN production over regular production using RESOLVERs. In JANCY this is used to dynamically initializer production: whether it would be token list on PASS1 or normal expression parsing on PASS2.