[axl.jnc] 'array' type modifier
most likely not gonna be used much, but for the sake of symmetry it has
to be there. it allows resolving semantic conflict of 'where-to-apply-
pointer'. it makes possible declaring pointers to arrays in one go, just
like 'function' modifiers allow declaring function pointers in one go:
int function* pf ();
int array* pa [10];