[axl.jnc] BUGFIX: line info of expression alias
when evaluating expression alias we we want to "lock" source position so
we DO NOT use source positions of alias tokens.

unfortunately, there are still situations when line info produces weird
jumps, but i have reason to believe it's LLVM fault in handling
debuglocs of the 'phi' nodes. hopefully it's gonna be fixed soon.