[axl_sys] fix: loop in ReadWriteLock::readLock is REQUIRED
also, added a similar loop in ReadWriteLock::writeLock -- both for symmetry and peace of mind (although I'm pretty sure it still will work without a loop)

also, separated writeCount into activeWriteCount (0 or 1) and queuedWriteCount -- for symmetry and simplified debugging (the actual state of a lock is non-obvious otherwise).

when everything is well-tested in IO Ninja, it's ok to do a micro-optimization by removing the loop from writeLock and merging activeWriteCount and queuedWriteCount into a single writeCount field.