[axl_sys] critical fix: sys::ReadWriteLock::writeLock/upgradeToWriteLock refused immediate write-lock if m_queuedWriteCount was non-zero (can lead to deadlock)
this can lead to deadlocks: let's say, we have two writer threads: (A) has obtained the lock, (B) is waiting. (A) releases the lock and immediately locks again. m_queuedWriteCount is non-zero, so we reset the write lock and start waiting. if it happens before (A) wakes up, we have a deadlock.