[axl_sys] critical fix: rare deadlock in ReadWriteLock; don't return bool from readLock/writeLock (spin-lock on wait failure)
originally, we did a reset of the read event before waiting on it. this was done to alternate readers and writers (and avoid writer starvation), but introduced an opportunity for a rare and subtle deadlock: if there were BOTH queued readers and queued writers, then a writer unlocks and signals the read event, then A NEW reader comes in (before any of the queued readers start waiting) and resets the read event -- deadlock. this almost never happened because a window of opportunity for this new reader was tiny (all this must have happened after unlock and before wait -- on all threads simultaneously)

the new design doesn't have this issue and it still guarantees alternating between readers and writers (but no guarantee on first-come-first-served for writers -- just like before).