[axl_enc] major update: emit original code units on UTF errors
a subtle, but very important update. imagine, a regex being applied to an *unintentionally* broken utf8 string like abc\x91def. if we replace broken utf8 sequences during decoding, we won't be able to search for patterns like this (originally, we bumped into this problem when trying to colorize \x91 in ioninja)